+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+**Version 2.0**
+- Renamed to [Steam Security Suite]
+- Rewrote the whole thing to only include phising protection.
+
 **Version 1.1**
 - Fixed all the CSS layout bugs.
 - Support for Enhanced Steam custom profile colors.